--- CityOfBinds/src/__init__.py
@@ -1,38 +1,0 @@
-from .triggers.trigger import Trigger, WASDTrigger
-from .triggers.mixin import TriggerMixin
-from .slash_commands.slashcommand import SlashCommand
-from .slash_commands.commandgroup import CommandGroup
-from .slash_commands.commandstemplate import CommandsTemplate
-from .slash_commands.power import Power
-from .binds.bind import Bind, WASDBind
-from .binds.bindtemplate import BindTemplate
-from .bind_file.comments import Comment, CommentBanner
-from .bind_file.bindfile import BindFile
-from .bind_file.bindfiletemplate import BindFileTemplate
-from .rotating_binds.rotatingbind import RotatingBind
-from ..utils.baseconverter import BaseConverter
-from ..utils.pathgenerator import PathGenerator
-from ..utils.Templates import templates as Templates
-from ..utils.Templates.pool import Pool
-
-__all__ = [
-    "Trigger",
-    "WASDTrigger",
-    "TriggerMixin",
-    "SlashCommand",
-    "CommandGroup",
-    "CommandsTemplate",
-    "Power",
-    "Bind",
-    "BindTemplate",
-    "WASDBind",
-    "Comment",
-    "CommentBanner",
-    "BindFile",
-    "BindFileTemplate",
-    "RotatingBind",
-    "BaseConverter",
-    "PathGenerator",
-    "Templates",
-    "Pool",
-]
