--- CityOfBinds/src/bind_graph_publisher/node.py
@@ -1,7 +1,0 @@
-from CityOfBinds.src.bind_file.bindfile import BindFile
-
-
-class BindFileNode:
-    def __init__(self, id: int, bind_file: BindFile):
-        self.id = id
-        self.bind_file = bind_file
