+++ CityOfBinds/src/content_managers/graph/__init__.py
@@ -1,0 +1,3 @@
+from .bind_file_graph import BindFileGraph
+
+__all__ = ["BindFileGraph"]
