+++ CityOfBinds/src/game/bind_file/comments/__init__.py
@@ -1,0 +1,4 @@
+from .comment import Comment
+from .comment_banner import CommentBanner
+
+__all__ = ["Comment", "CommentBanner"]
