+++ CityOfBinds/src/game/binds/__init__.py
@@ -1,0 +1,4 @@
+from .bind import Bind
+from .wasd_binds import WASDBind, iWASDBind
+
+__all__ = ["Bind", "WASDBind", "iWASDBind"]
