+++ CityOfBinds/src/game/rotating_binds/rotating_bind.py
@@ -1,0 +1,9 @@
+from .generic_rotating_binds import _GenericRotatingBind, _LoopTopology
+from ....utils.types import StrPath
+
+
+class RotatingBind(_LoopTopology, _GenericRotatingBind):
+    def __init__(self, is_silent: bool = True, absolute_path_links: bool = False):
+        _GenericRotatingBind.__init__(
+            self, is_silent=is_silent, absolute_path_links=absolute_path_links
+        )
