+++ CityOfBinds/src/game/utils/powers/__init__.py
@@ -1,0 +1,3 @@
+from .power import Power
+
+__all__ = ["Power"]
