+++ CityOfBinds/src/game/utils/slash_commands/__init__.py
@@ -1,0 +1,4 @@
+from .command_group import CommandGroup
+from .slash_command import SlashCommand
+
+__all__ = ["SlashCommand", "CommandGroup"]
