+++ CityOfBinds/src/game/utils/triggers/__init__.py
@@ -1,0 +1,5 @@
+from .trigger import Trigger
+from .wasd_trigger import WASDTrigger
+from .mixin import TriggerMixin
+
+__all__ = ["Trigger", "WASDTrigger", "TriggerMixin"]
