--- CityOfBinds/utils/__init__.py
@@ -1,4 +1,0 @@
-from .baseconverter import BaseConverter
-from .pathgenerator import PathGenerator
-
-__all__ = ["BaseConverter", "PathGenerator"]
