+++ CityOfBinds/utils/templates/constants.py
