+++ CityOfBinds/utils/types/__init__.py
@@ -1,0 +1,3 @@
+from .str_path import StrPath
+
+__all__ = ["StrPath"]
