+++ CityOfBinds/utils/types/str_path.py
@@ -1,0 +1,3 @@
+from pathlib import Path
+
+StrPath = str | Path
